Update: Rename `PostCardTableView` and make it public
